[v15] Add tag build support to Makefile (#38661)
* build: Generate centos7 archive when building linux-amd64

When building a linux-amd64 archive, make a copy of it with the
`centos7` tag as the linux-amd64 build works on centos7. We stopped
doing a centos7-specific build a while ago, but we still have the
archive on our releases page.

This helps unify the `release-amd64` and `release-amd64-centos7`
targets, which currently do the same thing except for the GitHub Actions
workflow that creates the centos7 archive only for the latter target.
This will allow us to get rid of that target as that latter target will
no longer be called when Drone is removed.

Also add a `-p` when making RELEASE_DIR as sometimes the parent
directory has not yet been created.

* build: Add `tag-publish` recipe to Makefile

Add a `tag-publish` recipe to the Makefile to start a workflow run of
the tag-publish workflow. The tag published is the current version tag
as defined by `$(VERSION)` in the Makefile. The version of the
`tag-publish` workflow used is that one tagged with `v$(VERSION)`.

* build: Add `tag-build` recipe to Makefile

Add a `tag-build` recipe to the Makefile to kick off a tag build on
GitHub Actions. Currently we cannot trigger the tag build in GitHub
Actions as the teleport repository is public so we are very careful with
credentials in it, and credentials are needed to run the `tag-build`
workflow in the `teleport.e` repository.

The new process for building a tag is to run `make update-tag` as usual,
but to then follow that with `make tag-build` to start the build.

* Update `e` ref

Updates include:
* Make the access list gRPC service use the cache. (gravitational/teleport.e#3544)
* Prevent overflow in sort functions provided to slices.SortFunc (#gravitational/teleport.e3549)
* Remove all quiet/silent flags from role checks. (#gravitational/teleport.e3540)
* [v15] fix: Lower bcrypt cost when testing (#gravitational/teleport.e3553)
* [v15] Add top-level workflow for building a tagged release (#gravitational/teleport.e3547)