Remove ResetSessionRecordingConfig with incorrect return type
The cluster config RPC service had a typo in the return type of
ResetSessionRecordingConfig. Since our proto linter doesn't allow
renaming, this removes the RPC entirely so that it can be added
with the correct signature in a follow up. This service was only
recently introduced, and is not exposed in any way with the Auth
grpc server so removal is a safe operation.