ci: Remove drone config file, turning off drone for teleport.e (#3565)
Remove `.drone.yml`, removing Drone from this teleport.e repo.

The only pipelines left are "push build" pipelines, which have limited
value right now. Much of what runs in the pipelines already runs on
GitHub Actions in the merge queue (Windows and Mac builds are done in
the merge queue), and any failures of push builds are not followed up
due to the noise and the fact that `master` of teleport.e does not
necessarily correspond to `master` of teleport, so failures are common.

A replacement for push builds is being devised.

Source-commit: gravitational/teleport.e@0e7258af9514d537c142dfb0bb161488f302da69
Source-pull-requests: gravitational/teleport.e#3565