[v14] AWS OIDC: List EC2, EICE and SG (#38443)
* AWS OIDC: List SecurityGroups service

This PR creates a new method on the AWS OIDC gRPC service that lists
SecurityGroups.

This is part of a refactor that moves the API calls behind the Auth
Service.

* AWS OIDC: List EC2 service (#38326)

This PR creates a new method on the AWS OIDC gRPC service that lists ec2
instances as Teleport Nodes.

This is part of a refactor that moves the API calls behind the Auth
Service.

* AWS OIDC: List EICE service (#38332)

* AWS OIDC: List EICE service

This PR creates a new method on the AWS OIDC gRPC service that lists ec2
instance connect endpoints.

This is part of a refactor that moves the API calls behind the Auth
Service.

* add required fields and aws api ref

* add test

* AWS OIDC: DeployService service (#38511)

* AWS OIDC: DeployService service

This PR creates a new method on the AWS OIDC gRPC service that deploys
an ECS Service.

This is part of a refactor that moves the API calls behind the Auth
Service.

* add rbac test and rename teleport config string

* revert e

* AWS OIDC: CreateEICE service (#38548)

This PR creates a new method on the AWS OIDC gRPC service that deploys
multiple EC2 Instance Connect Endpoints.

This is part of a refactor that moves the API calls behind the Auth
Service.