Prevent the debug flag from overriding trace logging
If the log severity in the config file was set to TRACE, which gets
parsed correctly when setting initial log level, AND teleport is
started with `-d/--debug` the log level was being downgraded from
the more verbose trace level to debug because the comparison was
case sensitive. By changing the comparison to be case insensitive
the more verbose option is now used even when `-d/--debug` is
applied.