ci: Remove drone(gen) from repo (#38759)
Remove `.drone.yml` and `dronegen` which generates it from the repo,
disabling drone on master. Remove references to drone in docs.

The remaining piplines in `.drone.yml` were the tag build and promote
pipelines which now run on GitHub Actions, the push build pipelines that
are currently of limited value and will be re-done differently with a
little thought, and the `teleport-docker-cron-ecr` pipeline which no
longer works as of v15 due to the removal of the legacy OCI images (the
teleport-labs stuff that the pipeline builds is to be removed shortly
too).