Add a Discovery Service section (#38848)
* Add a Discovery Service section

Closes #38741

Add a dedicated sidebar section for the Teleport Discovery Service.
Since the Discovery Service is a separate agent service from the SSH
Service, Kubernetes Service, etc., it makes sense to add a dedicated
section for this service. In the future, we can include configuration
guides, architecture guides, etc. in this section.

To create the new section, this change moves existing auto-discovery
guides from other sections of the docs.

* Respond to hugoShaka feedback

- Move the Kubernetes application discovery docs
- Edit the "Kubernetes Access" intro to include a link to Kubernetes
  auto-discovery