Set maximum message size for gRPC in accessgraph (#3557)
* Set maximum message size for gRPC in accessgraph

This commit sets and enables a maximum message size of 50MB for gRPC bindings within the accessgraph. The implementation applies not only to query functions but also the "GetFile" method ensuring efficient message handling and protection against potential DDoS attacks.

* Reduce max message size and reorganize imports

The maximum size for GRPC messages in the Access Graph service was reduced from 50MB to 20MB to limit payload size. The order of the import statements was also rearranged as per Go conventions.

Source-commit: gravitational/teleport.e@2db1971a802a96ad6e997a3c285a4c46c6aa1eae
Source-pull-requests: gravitational/teleport.e#3557