Register cluster configuration service (#38734)
Updates the auth gRPC server to include the new clusterconfigv1
service. The api client was also updated to allow interacting
with the new service(cluster audit config not included yet). Nothing
has yet been converted to use the new service, that will come in a
follow up.

The RPC service proto was also updated to fix a typo that was
reverted in https://github.com/gravitational/teleport/pull/38736.