ci: Add workflow to build all our buildboxes (#3617)
Add a workflow to build all our buildboxes, triggered by cron and
manually. We cannot build the arm64 buildboxes from the OSS Teleport
repo as they do not have access to the self-hosted runners we have that
are arm64. The arm64 buildboxes need to be built on an arm64 host.

The `schedule` trigger builds all the buildboxes daily. This should be
sufficient to keep them up-to-date that when a tag-build runs, it should
require no rebuilding of the buildboxes.

The OSS teleport repo has a `push` trigger that fires if any of the
source files of the buildboxes change on merge to `master` /
`branch/**`. This build the amd64 buildboxes but cannot build the arm64
buildboxes as that requires our self-hosted runners which are not
available to the OSS repo. For that we have a `workflow_dispatch`
trigger on this workflow to manually kick off the arm64 buildbox builds
until we can figure an automated trigger.

The branch the buildboxes are built for is determined by the GitHub ref
of the triggering event. For development of this workflow, a `branch`
input parameter can target a different branch. This is often needed as
the development branch in teleport.e likely won't exist in teleport and
it is easier to test the release branches without needing to backport a
PR right away.

Source-commit: gravitational/teleport.e@ac49d2a75014d44a66fc6bf538329e8be4f1f030
Source-pull-requests: gravitational/teleport.e#3617