[v13] build: Sign mac artifacts using key name not hash
Tidy up the mac signing makefile magic to sign using the key name and
not the key hash. Since we look up the key has by the name, there's no
benefit to using the hash as all the signing commands will take the key
name as well. This removes a bunch of fragility around exporting
variables, make errors and running the key lookup commands only on a mac
that it leaves it quite a bit simpler.

The `CSC_NAME` variable used by electron-builder to build Teleport
Connect does not want the full key name - it wants the purpose part
stripped off the front (`Developer ID Application:`), so we do that. The
other signing commands take the full key name.

This also simplifies the `NOTARIZE_BINARIES` set of commands, allowing
them to be broken up removing the `&&` conjunction. This makes it clear
exactly what command fails if there is a failure as make runs them
one-by-one instead of all at once. It also allows the Makefile in
teleport.e a little simpler as it no longer needs to run the
notarization from the parent directory.