[v15] Consolidate webasset bundle to single chunks (#38989)
* Prevent code splitting

This will prevent the react bundle from being code split. This means
that our entire javascript bundle will be a single file as well as our
css file. It also removes hashing from the included webassets. The reason
for this change is to allow different versioned proxies to coexist behind a
load balancer. If we remove code splitting, then the only time our bundle
will load is on initial hit, and the web client will then have the entire
bundle.

This also removes React.lazy. React.lazy wrapped components in a promise
to allow for dynamic imports during runtime. This is no longer needed.

Lastly, to clean up the default exports used by react lazy, this removes
(most) default exports for feature components. Assist and telemetry-boot
are not included in the javascript bundle

* Fix the import on the account settings page (#39015)

* Fix auth connectors import

---------

Co-authored-by: Bartosz Leper <bartosz.leper@goteleport.com>