Improve Linux installation instructions (#36596)
* Improve Linux installation instructions

Closes #32387

In the Installation page, the instructions for installing Teleport on
Linux reuses the partial we use for how-to guides, `install-linux.mdx`.
This is inappropriate for an installation reference, since the partial
is not intended for users interested in a specific package manager. This
change provides a comprehensive description of our package repositories
for Teleport installations on Linux.

Since the Installation page is a common entrypoint for users looking to
get started with Teleport, this change also adds a section to the Linux
installation instructions with recommended pages to follow for setting
up Teleport.

Also, since we have a script available form the corporate Teleport site
that automatically chooses a package manager and configures it to target
a Teleport package repo, this change adds instructions to use this.

I have organized the Linux installation section to place H3 sections
from less work/configuration to more work/configuration, from the
one-line installation script to package manager-specific instructions
and, finally, general information about our package repositories.

* Respond to r0mant feedback

- Remove the version guidance for Cloud users and make the
  recommendation for them to enable automatic agent upgrades more
  prominent.

- Add instructions for installing `teleport-ent-updater`.

- Add instructions for changing the Teleport version to `cloud` for
  Team/Enterprise Cloud in package manager examples. Reorganize the
  package manager instructions so there is a single tabbed example of
  setting the version. To account for this, refactor the subsections of
  the "Linux" section so there is a single H3 for TAR and a single H3
  for package repositories.

* Respond to additional r0mant feedback

- Make it clear throughout the section that we recommend using the
  one-line installation script, which install the automatic updater for
  Teleport Cloud users.

- Condense package repository and TAR archive H3s. By simplifying these
  sections, we can make the instructions to use the updater on Cloud
  more prominent.

- Warn against using TAR archives with Teleport Cloud, and remove
  TabItems for these editions in the TAR archive section.

- Incorporate instructions introduced by #38401 to query the Proxy
  Service for the latest compatible version.

* Respond to r0mant feedback

- Remove Teleport Team mention
- Add "(self-hosted only)" to the TAR subsection
- Remove a Tabs component in the TAR section so readers aren't tempted
  to interact with the component before reading the initial body text.

* Fix linter errors