release: Add new changelog generation script
Add a new `changelog.sh` script that has a couple of nice features over
the old one:
* Can run from the base release branch without args and it generates a
  changelog since the last release. The base branch and tag are still
  overrideable though if you need to
* Handles multiple changelog entries on one PR which can occasionally
  happen.
* Cleans up the changelog entries a little (trailing period, cleans
  whitespace).
* Marks entries that did not have a changelog where the PR title was
  used. These need to be checked and likely re-written.
* Extracts enterprise changelogs.

The primary use case is to be able to run and re-run it easily as you
tweak the changelog entries on the PRs so you get a final generation run
that needs no changes to the changelog itself. It is also easy to update
the changelog if new PRs have been merged into the base branch before
the release PR was able to be merged.