Lock AWS Sync into a single discovery service
This PR prevents multiple `discovery_services` running with the same
configuration to race against each other and push new resources to
Access Graph service.

This PR also introduces the notion of features to identify the list of
supported resources by the Access Graph service to which the service is
connected to.

Paired with https://github.com/gravitational/access-graph/pull/591

Signed-off-by: Tiago Silva <tiago.silva@goteleport.com>