Update eslint-plugin-jest to 27.9.0
https://github.com/jest-community/eslint-plugin-jest/blob/main/CHANGELOG.md#2600-2022-01-24
https://github.com/jest-community/eslint-plugin-jest/blob/main/CHANGELOG.md#2700-2022-08-28

I switched plugin:jest/all to plugin:jest/recommended, since /all just
activates every rule in the package. This might have been fine when this
package was added a couple of years ago, but it doesn't make sense anymore.

27.0.0 requires parserOptions.project to be set, but this returned an error:

    0:0  error  Parsing error: ESLint was configured to run on
    `<tsconfigRootDir>/web/packages/teleport/vite.config.mts` using `parserOptions.project`:
    /users/rav/src/teleport/tsconfig.json
    However, that TSConfig does not include this file. Either:
    - Change ESLint's list of included files to not include this file
    - Change that TSConfig to include this file
    - Create a new TSConfig that includes this file and include it in your parserOptions.project

I updated tsconfig.node.json so that it applies only to .mts files, as
I understand that was the original intention.