ci: Specify FIPS dockerfile when rebuilding distroless OCI images (#3672)
When doing the nightly rebuild of the distroless OCI images, specify the
Dockerfile to use so that the FIPS dockerfile is used when building FIPS
images. This change was originally implemented in cd2a1e59e for building
the images, but was not made to the rebuild workflow, so that has been
failing since the changes were backported to release branches.

Source-commit: gravitational/teleport.e@d426ca521b4215209d4dc45a3547f169ced692c9
Source-pull-requests: gravitational/teleport.e#3672