Migrate from teleport.Component to teleport.CompoundComponent
This is the first step in a series of changes that will result in
teleport.Component replacing trace.Component. Since
teleport.Component already exists as a function, which is also
consumed in teleport.e, a duplicate teleport.CompoundComponent was
added. The original function is left in place for now, until the
references in teleport.e can be converted, at which time the
teleport.Component function can be replaced by the constant.