Edit recording playback instructions in tsh guide
Closes #26521

Add `tsh play` instructions for databases by editing the `tsh` guide in
the "Connect your Client" section. The section currently assumes that
the user is playing sessions via PTY, so this change includes the
possibility of using the `yaml` and `json` values of the `--format`
flag.

This change also restructures the section of the `tsh` guide related to
session playback to accommodate the new instructions. It also mentions
Windows desktops for completeness.

Finally, the current version of the guide includes a brief note about
`tsh recordings ls` flags that is not complete enough to be useful. This
change removes the note and links to the `tsh` CLI reference at the end
of the guide for further information about flags and arguments.