Access list owners can once again edit members when they are members. (#3684)
* Access list owners can once again edit members when they are members.

A while ago, a fix to ensure that access list owners were able to edit access
list members when the owner was, themselves, a member. This regressed
recently, it seems when ensuring users could specific fields in an Okta access
list.  This fix allows owners to edit access list members members again when
they are members of the access list.

* Unit tests for populate members, verify joined is preserved.

* Test code cleanup, add in owner modify OK test.

* Fix up tests.

* Update lib/accesslist/service_test.go

Co-authored-by: Alan Parra <alan.parra@goteleport.com>

---------

Co-authored-by: Alan Parra <alan.parra@goteleport.com>

Source-commit: gravitational/teleport.e@94fa25b93ad0cdc7fb2c0f9e7b7e816b2307abd3
Source-pull-requests: gravitational/teleport.e#3684