Fix ListUnifiedResources RBAC
RBAC works slightly different with the unified resources API so
that if a user requests multiple resource kinds, and is only
permitted to access a subset of them, only the allowed resources
are returned instead of a trace.AccessDenied error. However, the
logic for this had a slight omission if the user has access to none
of the requested kinds. Prior to this change, in that scenario the
API would return nil, nil instead of nil, trace.AccessDenied.