[v15] AWS OIDC Integration: use S3 bucket as IdP issuer (#39489)
* AWS OIDC: Accept custom issuers (#38785)

* AWS OIDC: Accept custom issuers

This PR adds a new field into the AWS OIDC fields:
- issuer

This is the issuer that was configured in AWS.
It is used by teleport to set the `issuer/iss` field when generating the
JWT.
This way we'll be able to set a custom issuer.
As an example, we could set this to be a public S3 bucket which doesn't
suffer from the thumbprint validation issuer.

* ensure issuer is a valid https url

* use s3 uri instead of any url for issuer

* typo and require s3 fields

* add missing s3 location when creating integration

* improve error messages

* AWS OIDC IdP Configuration: support S3 buckets as provider (#39026)

This PR adds support for setting the AWS OIDC Integration with an S3
bucket as issuer.

* AWS OIDC: Require S3 for configure IdP Script (#39113)

There are two new required fields for generating the configure IdP
script:
- s3Bucket
- s3Prefix

This must form a valid URI when joining them:
s3://<s3Bucket>/<s3Prefix>