ci: Don't build webassets when building universal binaries (#3715)
* ci: Don't build webassets when building universal binaries

Remove the line from the `release:` recipe that builds the webassets.
This is the wrong place for it as `tool/teleport` already has a
dependency on the building the webassets, and they do not need to be
built if `tool/teleport` is not being built from Go code.

This speeds up the build of the universal binaries as they do not build
`tool/teleport` from source but use `lipo` to combine the amd64 and
arm64 binaries into a universal binary.

This also allows us to remove the step that installs rust and builds
wasm-pack from the `build-mac-universal` workflow. Combined with not
building the webassets, this cuts about 4m30s from the build time, about
25%.

* ci: Remove python install from build-mac-universal workflow

Remove the "Set up Python" step from the `build-mac-universal` workflow
as we have updated `node-gyp` to v10. According to the linked issue,
that updates `gyp` to v0.16.1` and removes the need to install python.

Resolved-by: https://github.com/gravitational/teleport/pull/35776
Link: https://github.com/nodejs/node-gyp/issues/2869

Source-commit: gravitational/teleport.e@67b61f002d82488a606c0ac07242f241a0280ed8
Source-pull-requests: gravitational/teleport.e#3715