Make reconciler use map keys as resource identity (#39547)
Prior to this change, the reconciler required that the input
map keys be equal to the value returned by calling GetName() on
the map value.

This patch removes this requirement, and uses the map key directly as
the identity of a resource. This allows the reconciler to be used
more flexibly in situatons where the resource name may not be a suitable
identity.

The motivating context is the AccessListMember reconciler. Using the
AccessListMember (i.e. the member username) is fine for reconciling the
users for a single AccessList, as the map key and resource name are the
same. This falls apart when the reconciler is managing reconciliation
across multiple access lists at once. In this case, the map keys are
typically a generated string like `${ACCESS_LIST}/${MEMBER_USER_NAME}`.

In this case, the reconciler will attempt to match existing and updated
resources using the `AccessListMember` name (i.e. the member username)
as the `AccessListMember` identity. This identity is obviously not present
in the updated resources map, which expects the AccessList+Username pair
and so the reconciler promptly deletes the associated resource (which is
then promptly re-created by the reconciler while it processes processing
any new resources).

This patch sidesteps all that by only using the input map keys as the
resource identity.