dependabot - Temporary removal of groups config
We currently are failing to get Dependabot updates for `/`.  After discussing with GitHub support, they suggested for debugging we temporarily remove the `groups` config, apply updated, then re-apply the `groups` configuration.

This is a temporary disable of the beta feature to hope to get Dependabot into a working state again.