ci: Make publish-os-packages dispatchable (#3704)
Change the type of the `environment` input from `environment` to
`string` for `workflow_dispatch` as the actual value required is not a
real environment name, but a "pseudo-environment" name. Update the
description of the input parameter to describe this.

Without this change, the workflow cannot be dispatched as the valid
input values as required by the workflow itself is not valid to GitHub
when using `environment` as the type.

Add a `permission:` block for what permissions this workflow needs to
run. When this workflow is triggered via a `workflow_call`, the
`tag-publish` workflow supplies these permissions, but when dispatched
directly with `workflow_dispatch`, this workflow needs to request the
permissions itself.

Source-commit: gravitational/teleport.e@65e2692748b62e38f27475049bcde55488e1a245
Source-pull-requests: gravitational/teleport.e#3704