Rename Dockerfile-connect to Dockerfile-node (#39483)
This aligns branch/v13 with master and branch/v15 WRT the name of the
buildbox with node.

This was originally done to facilitate building webassets by adding Rust
and wasm-pack to the buildbox, so renaming the buildox made sense in
that context. Rust and wasm-pack are not used on branch/v13 so this was
never backported, but we want to keep the build across branches aligned
as much as possible to make other backports easier.

Co-authored-by: Isaiah Becker-Mayer <isaiah@goteleport.com>