Document how to import resources into Terraform
Closes #9720

Add a guide to importing Teleport resources into Terraform. While it is
mostly straightforward to add an `import` block, determining the ID to
use for each resource requires some planning, since the ID depends on
the resource type.

This change also edits the Terraform provider guide, since testing out
this guide was required to test out the resource import workflow:

- Minor style and readability tweaks.
- Remove the `-f` flag in `tctl create`: This way a user doesn't
  accidentally override an existing `terraform` role
- Add a `tbot` data directory in an unprivileged path in the working
  directory of the setup script. This way, we don't need to give
  `teleport` ownership of the default Machine ID output directory.
- Fix identity file paths. These should point to a file called
  `identity` in the `terraform-identity` directory.
- Edit the "Next steps" section to point to the resource import guide.