build: Add cloud-only tag-build/publish option
Update the `tag-build` and `tag-publish` recipes to call the cloud-only
variants of the workflows if the `VERSION` variable contains `-cloud.`
or `-dev.cloud.`, or if `CLOUD_ONLY` is set on the `make` command line.

The cloud-only workflows only build the Linux tarballs and OS packages
and the distroless OCI images. This allows for a faster release build
for cloud-only releases which does not need any of the other release
artifacts. As a pre-release version, it will ultimately be superseded by
a regular release that does contain all the release artifacts.