Add a vale rule and test suite
In how-to guides, ensure that the first H2 section is called "How it
works". This section is intended to include architectural information.
Architecture is often a point of confusion among Teleport users, so
adding this rule ensures that how-to guides include an attempt at an
architectural discussion.

This is intended as a complement to any dedicated architectural docs.
Since most users visit the docs through organic search, we can't
guarantee that a user visiting a guide will have discovered the relevant
architectural information.

Since the vale rule includes some logic, this change also adds a test
suite for vale rules.