[v15] Enforce optimistic locking for AuthPreferences, ClusterNetworkingConfig, SessionRecordingConfig (#39785)
* Register cluster configuration service (#38734)

Updates the auth gRPC server to include the new clusterconfigv1
service. The api client was also updated to allow interacting
with the new service(cluster audit config not included yet). Nothing
has yet been converted to use the new service, that will come in a
follow up.

The RPC service proto was also updated to fix a typo that was
reverted in https://github.com/gravitational/teleport/pull/38736.

* Convert SetClusterNetworkingConfig usages to new Create/Update/Upsert API (#39219)

Migrates callers off the deprecated API and onto the new API so that
ClusterNetworkingConfig can be protected via optimistic locking when
appropriate.

* Convert SetAuthPreferences usages to new Create/Update/Upsert API (#38895)

Migrates callers off the deprecated API and onto the new API so that
AuthPreferences can be protected via optimistic locking when appropriate.

* Convert SetSessionRecordingConfig usages to new Create/Update/Upsert API (#39249)

Migrates callers off the deprecated API and onto the new API so
that SessionRecordingConfig can be protected via optimistic locking
when appropriate.