Add tctl sso configure and tctl sso test support for OIDC and SAML (#39899)
* Add tctl sso configure and tctl sso test support for OIDC and SAML

Migrates the tctl sso configure and tctl sso test commands for
SAML and OIDC connectors from tctl enterprise to OSS. This will
allow the enterprise tctl binary to be removed.

* Guard CreateOIDCAuthRequest and CreateSAMLAuthRequest

Prevent the RPCs from being called in the open source version of
teleport. Now that the tctl sso test commands are in OSS we want
to ensure that the commands will only function if communicating
with an enterprise Auth server.

* fix: license

* update e ref