Enforce a "How it works" section in how-to guides
Architecture is often a point of confusion among Teleport users. When
following how-to guides, it is important for users to understand how the
components of a setup fit together so they can (a) form a mental model
of where each step first in an overall procedure and (b) troubleshoot
more easily if they encounter the unexpected.

This change ensures that the first H2 section of a how-to guide is
called "How it works" by adding a vale
[script](https://vale.sh/docs/topics/styles/#script) rule. As the error
message associated with the rule indicates, the "How it works" section
is intended to include architectural information.

The "How it works" section is intended as a complement to any dedicated
architectural docs. Since most users visit the docs through organic
search, we can't guarantee that a user visiting a guide will have
discovered the relevant architectural information.