Update slog json handler to output error text (#39991)
The default slog json handler attempts to format attributes with
json.Marshal instead of a string representation, which resulted in
errors being output as json objects instead of just the message.

```
"error":{"error":{"message":"Unable to communicate with the Teleport Auth Service"}}
```

In order to restore json output to contain only the error message
instead of the entire object special handling is done in our slog
handler wrapper that replaces the error object with the message.