Simplify `install-linux.mdx`
Closes #38933

Simplify the instructions in the `install-linux.mdx` partial by
including instructions for the one-line installation script instead of
having the user choose the Teleport edition from a dropdown menu, then
choose a package manager. The one-line installation script chooses a
package manager automatically.

Since we only include `install-linux.mdx` in guides to installing
Teleport on agent nodes, instruct the user to get the version of either
the automatic updater (if automatic updates are enabled) or the Teleport
cluster.

This approach has the advantage of not using tabs, so it creates less
noise when authors include this partial in a `Tabs` component.

Also include a link to our installation guide at the end of the partial
for readers who want more information on our package repos.