Add reissuing user certificates to the ClusterClient
Updating the ClusterClient to support generating new user certs
covers some of the exiting missing functionality that is required
by some components to migrate fully off of the deprecated ProxyClient.
The ClusterClient implementation is almost a verbatim copy from
the ProxyClient.

The ProxyClient method was not removed because Connect does not
use the ClusterClient yet. That will be addressed after the
ClusterClient has complete feature parity with the ProxyClient.

Updates #38239.