Replace UnixShellQuote with `shsprintf.EscapeDefaultContext` from github.com/google/safetext (#40046) (#40060)
This change is a change in how we are handling the untrusted input.  Instead of adopting a strategy of quoting (which was incomplete), this instead escapes any special characters.

Although github.com/google/safetext is new, and without any current tags for us to reference, I believe it would be better to reference their implementation than to update the UnixShellQuote to handle the newly identified cases.