Remove lingering Drone references
Drone is dead and gone. This change removes a few of its ghosts so
that others won't trip on them.

There are a couple more that aren't as easy to remove such as relcli
depending on DRONE_REPO and DRONE_TAG, but we'll loop back to pick them
up in time.

Contributes to https://github.com/gravitational/SecOps/issues/481