ci: Always publish operator OCI images even if tags exist (#3802)
Fix the teleport operator OCI publishing workflow to always publish
images even if the tag exists. This is needed as we always update some
tags with the MAJOR and MAJOR.MINOR version tags and if we skip the
because they exist, we do not provide the latest version with these
tags.

Source-commit: gravitational/teleport.e@c532b2fdc575f64228e4b999d8cd379e3db35607
Source-pull-requests: gravitational/teleport.e#3802