Add a way to configure the `RunWhileLocked.RunWhileLockedRetryInteral` for tests (#40083)
* Add a way to configure the `RunWhileLocked.RunWhileLockedRetryInterval` for tests

This PR adds a way to bypass the
`cfg.Backend.Clock().After(cfg.RetryInterval)` call within
`backend.AcquireLock` when using the fake clock.

When multiple parallel calls try to acquire the same lock,
`backend.AcquireLock` tries to sleep before retrying. When using the
fake clock this operation isn't reliable enough when using
`clock.Advance` so we are adding a way to configure it with negative
values - i.e. retry immidately - for testing purposes.

Signed-off-by: Tiago Silva <tiago.silva@goteleport.com>

* update docs

* drop testing.Testing usage

---------

Signed-off-by: Tiago Silva <tiago.silva@goteleport.com>