Refactors rust lint out into its own job which runs on a bare GH runner
Also adds -s to silence all make commands in workflows which are used as
some form of output in the workflow script. This is to prevent the
workflow from failing due to noise from the make command (such as when
make tells you it's entering a directory, etc).

Also updates to latest e-ref